+++ backend/src/modules/users/interfaces/applications/update.user.service.interface.ts
@@ -1,6 +1,6 @@
 import { LeanDocument } from 'mongoose';
 import UpdateUserDto from '../../dto/update.user.dto';
-import { UserDocument } from '../../schemas/user.schema';
+import User, { UserDocument } from '../../entities/user.schema';
 
 export interface UpdateUserApplication {
 	setCurrentRefreshToken(
@@ -12,7 +12,7 @@ export interface UpdateUserApplication {
 		userEmail: string,
 		newPassword: string,
 		newPasswordConf: string
-	): Promise<UserDocument | null>;
+	): Promise<User | null>;
 
 	checkEmail(token: string): Promise<string>;
