+++ backend/src/modules/users/interfaces/services/update.user.service.interface.ts
@@ -1,15 +1,15 @@
 import { LeanDocument } from 'mongoose';
 import UpdateUserDto from '../../dto/update.user.dto';
-import { UserDocument } from '../../schemas/user.schema';
+import User, { UserDocument } from '../../entities/user.schema';
 
 export interface UpdateUserService {
-	setCurrentRefreshToken(refreshToken: string, userId: string): Promise<UserDocument | null>;
+	setCurrentRefreshToken(refreshToken: string, userId: string): Promise<User | null>;
 
 	setPassword(
 		userEmail: string,
 		newPassword: string,
 		newPasswordConf: string
-	): Promise<UserDocument | null>;
+	): Promise<User | null>;
 
 	checkEmail(token: string): Promise<string>;
